Subscribe to multiple channels
I wanted this package to be a full function useful package for react
components to talk to ActionCable.  The package was restricted to
working with a single subscription but I refactored it to be able to
allow for multiple subscriptions and custom cable endpoint.  Also added
a list of actions to make it super simple to work with ActionCable.